refactor breakdown components into separate file
